Fix some Qt6 migration bits that were missed
I did my own port to Qt6 on the stable branch and noticed I'd made some
changes that were missing here.

It was easy to apply these to the beta branch, but the master branch
will require more work.

Yes, a lot of these are just comments, but we may as well fix those too.